refactors footer.tsx (#394)
